requirements.txt
Pillow==10.1.0
opencv-python==4.9.0.80
numpy==1.26.2
pandas==2.0.3
requests==2.31.0
urllib3==1.26.6